pdfwrite enhancement : Final DSC changes
Fix a compiler warning, make ProduceDSC=true the default condition so that we create
DSC-compliant PostScript unless told otherwise.

Update the documentation for pswrite and ps2write in devices.htm and Ps2ps2.htm. Note
that pswrite is now deprecated.

Modify the pdf2ps and ps2ps scripts to use ps2write instead of pswrite.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11952 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/base/gdevpdfu.c
@@ -384,7 +384,7 @@ int ps2write_dsc_header(gx_device_pdf * pdev, int pages)
 	sprintf(BBox, "%%%%BoundingBox: 0 0 %d %d\n", width, height);
 	stream_write(s, (byte *)BBox, strlen(BBox));
 	pdf_get_docinfo_item(pdev, "/CreationDate", cre_date_time, sizeof(cre_date_time));
-	sprintf(BBox, "%%%%Creator: %s %d (%s)\n", gs_product, gs_revision,
+	sprintf(BBox, "%%%%Creator: %s %d (%s)\n", gs_product, (int)gs_revision,
 	        pdev->dname);
 	stream_write(s, (byte *)BBox, strlen(BBox));
 	stream_puts(s, "%%LanguageLevel: 2\n");
